Refactor endpoints.rs. Split it into multiple files
