+++ src/api/pokemon.rs
@@ -1,0 +1,36 @@
+use crate::models::pokemon::Pokemon;
+
+use rocket::State;
+use rocket::serde::json::Json;
+
+use surrealdb::Surreal;
+use surrealdb::engine::remote::ws::Client;
+
+#[get("/pokemon/get/<id>")]
+pub async fn get(id: u64, db: &State<Surreal<Client>>) -> Option<Json<Pokemon>> {
+    let pokemon: Option<Pokemon> = match db.select(("pokemon", id)).await {
+        Ok(p) => p,
+        Err(e) => {
+            println!("{}", e);
+            return None;
+        }
+    };
+
+    match pokemon {
+        Some(p) => Some(Json(p)),
+        None => None,
+    }
+}
+
+#[get("/pokemon/get")]
+pub async fn list(db: &State<Surreal<Client>>) -> Json<Vec<Pokemon>> {
+    let pokemon: Vec<Pokemon> = match db.select("pokemon").await {
+        Ok(p) => p,
+        Err(e) => {
+            println!("{}", e);
+            Vec::new()
+        }
+    };
+
+    Json(pokemon)
+}
